엔티티 수정

+++ src/main/java/com/ysshop/shop/entity/Cart.java
@@ -13,7 +13,8 @@ import lombok.ToString;
 public class Cart extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer cartId;
+    private Long cartId;
+
     private Integer quantity;
 
     @ManyToOne
